UC-14 refactor: refactor code
